Refactor code to support react router
