+++ package.json
@@ -10,6 +10,7 @@
     "aws-amplify": "^3.0.14",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "styled-components": "^5.1.1"
   },
